Rename `ClientConstants` to `ClientInternal`
